Add auto-generated typings for RTDB
